Fix for Dashboard

+++ StudentPortal/src/main/java/com/protean/student/StudentPortal/repository/EventsDetailsRepository.java
@@ -14,6 +14,6 @@ public interface EventsDetailsRepository extends JpaRepository<EventDetails, Lon
 	 * EventDetails findByEventid(Integer eventId);
 	 */
 	
-	EventDetails findByEventid(Integer eventId);
+	EventDetails findByEventid(Long eventId);
 
 }
